requirements.txt
Flask
gunicorn
openai>=1.3.0
twilio
psycopg2-binary
